front_end/requirements/base.txt
Django==1.5.4
Markdown==2.3.1
Pillow==2.5.3
South==0.7.6
Unipath==1.0
amqp==1.3.3
amqplib==1.0.2
anyjson==0.3.3
billiard==3.3.0.13
boto==2.8.0
celery==3.1.7
distribute==0.6.34
django-appconf==0.6
django-compressor==1.3
django-extensions==1.0.3
django-redis==3.4.0
tornado-redis==2.4.15
sockjs-tornado==1.0.0
gunicorn==0.17.2
ipython==0.13.1
jsonfield==0.9.4
kombu==3.0.8
numpy==1.7.1
pika==0.9.12
psycopg2==2.4.6
pycurl==7.19.0
python-dateutil==2.1
pytz==2012j
requests==1.2.3
simple-salesforce==0.51
six==1.4.1
stripe==1.7.9
validictory==0.9.3
djangorestframework==2.3.12
django-rest-swagger==0.1.14
django-filter==0.7
django-cors-headers==0.12
createsend==4.0.0
